Split "font finding" functions into separate file, clarify some names, fix bug introduced by nesting fonts dir
